requirements.txt
# FakeNews + Deepfake Detector Requirements

# Core ML/AI libraries
transformers>=4.30.0
torch>=2.0.0
torchvision>=0.15.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0

# Web Scraping
newspaper3k>=0.2.8
lxml>=4.9.0
lxml_html_clean>=0.1.0
beautifulsoup4>=4.12.0

# Web Interface
streamlit>=1.28.0

# Utilities
requests>=2.31.0
python-dateutil>=2.8.2
python-dotenv>=1.0.0
newsapi-python>=0.2.7

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Optional: For better performance
# accelerate>=0.20.0  # Uncomment for faster model loading
# sentencepiece>=0.1.99  # Uncomment if models require it
